+++ src/ledger/LedgerMaster.cpp
@@ -100,17 +100,17 @@ Database &LedgerMaster::getDatabase()
     return mApp.getDatabase();
 }
 
-int32_t LedgerMaster::getTxFee()
+int64_t LedgerMaster::getTxFee()
 {
     return mCurrentLedger->mHeader.baseFee; 
 }
 
-int64_t LedgerMaster::getMinBalance(int32_t ownerCount)
+int64_t LedgerMaster::getMinBalance(uint32_t ownerCount)
 {
     return (2 + ownerCount) * mCurrentLedger->mHeader.baseReserve;
 }
 
-int64_t LedgerMaster::getLedgerNum()
+uint64_t LedgerMaster::getLedgerNum()
 {
     return mCurrentLedger->mHeader.ledgerSeq;
 }
